+++ railties/lib/rails/application/configuration.rb
@@ -4,7 +4,6 @@ require "rails/engine/configuration"
 require "rails/source_annotation_extractor"
 
 require "active_support/deprecation"
-require "active_support/core_ext/string/strip" # for strip_heredoc
 
 module Rails
   class Application
@@ -54,24 +53,6 @@ module Rails
         @debug_exception_response_format = nil
         @x                               = Custom.new
         @enable_dependency_loading       = false
-      end
-
-      def serve_static_files
-        ActiveSupport::Deprecation.warn <<-eow.strip_heredoc
-          `config.serve_static_files` is deprecated and will be removed in Rails 5.1.
-          Please use `config.public_file_server.enabled` instead.
-        eow
-
-        @public_file_server.enabled
-      end
-
-      def serve_static_files=(value)
-        ActiveSupport::Deprecation.warn <<-eow.strip_heredoc
-          `config.serve_static_files` is deprecated and will be removed in Rails 5.1.
-          Please use `config.public_file_server.enabled = #{value}` instead.
-        eow
-
-        @public_file_server.enabled = value
       end
 
       def encoding=(value)